+++ lib/flutter_automation.dart
@@ -6,14 +6,20 @@ import 'package:args/args.dart';
 import './firebase_auth.dart' as firebase_auth;
 import './google_maps.dart' as google_maps;
 import './android_signing.dart' as android_sign;
+import './firestore_crud.dart' as firestore_crud;
 
 void decipherScript(List<String> arguments) {
   var parser = ArgParser(allowTrailingOptions: true);
-  parser.addFlag('help',abbr: 'h', negatable: false, help: "Usage help");
-  parser.addFlag('firebase-auth',abbr: 'f', help: "Adds firebase authentication", negatable: false);
-  parser.addFlag('google-maps',abbr: 'g', help: "Adds google maps", negatable: false);
+  parser.addFlag('help', abbr: 'h', negatable: false, help: "Usage help");
+  parser.addFlag('firebase-auth',
+      abbr: 'f', help: "Adds firebase authentication", negatable: false);
+  parser.addFlag('firestore-crud',
+      abbr: 'c', help: "Adds firestore CRUD boilerplate", negatable: false);
+  parser.addFlag('google-maps',
+      abbr: 'g', help: "Adds google maps", negatable: false);
 
-  parser.addFlag("android-sign",abbr: 's', help: "Setups android signing config");
+  parser.addFlag("android-sign",
+      abbr: 's', help: "Setups android signing config", negatable: false);
 
   var argResults = parser.parse(arguments);
 
@@ -23,15 +29,19 @@ void decipherScript(List<String> arguments) {
     return;
   }
 
-  if(argResults['firebase-auth']) {
+  if (argResults['firebase-auth']) {
     firebase_auth.firebaseAuth();
   }
 
-  if(argResults['google-maps']) {
+  if (argResults['google-maps']) {
     google_maps.googleMaps();
   }
 
-  if(argResults['android-sign']) {
+  if (argResults['android-sign']) {
     android_sign.androidSign();
+  }
+
+  if (argResults['firestore-crud']) {
+    firestore_crud.firestoreCrud();
   }
 }